Move the object segement out of the type file.
