+++ source/world_builder/features/fault.cc
@@ -26,6 +26,7 @@
 #include "world_builder/types/double.h"
 #include "world_builder/types/object.h"
 #include "world_builder/types/point.h"
+#include "world_builder/types/segment.h"
 #include "world_builder/types/unsigned_int.h"
 #include "world_builder/world.h"
 #include <algorithm>
