+++ source/world_builder/features/subducting_plate.cc
@@ -25,6 +25,7 @@
 #include "world_builder/types/double.h"
 #include "world_builder/types/object.h"
 #include "world_builder/types/point.h"
+#include "world_builder/types/segment.h"
 #include "world_builder/types/unsigned_int.h"
 #include "world_builder/world.h"
 #include <algorithm>
