+++ source/world_builder/types/segment.cc
@@ -167,82 +167,5 @@ namespace WorldBuilder
 
     }
   } // namespace Types
-
-  namespace Objects
-  {
-    // todo update function
-    template<class A, class B, class C>
-    Segment<A,B,C>::Segment(const double default_length_,
-                            const WorldBuilder::Point<2> &default_thickness_,
-                            const WorldBuilder::Point<2> &default_top_truncation_,
-                            const WorldBuilder::Point<2> &default_angle_,
-                            const std::vector<std::shared_ptr<A> > temperature_systems_,
-                            const std::vector<std::shared_ptr<B> > composition_systems_,
-                            const std::vector<std::shared_ptr<C> > grains_systems_)
-      :
-      value_length(default_length_),
-      default_length(default_length_),
-      value_thickness(default_thickness_),
-      value_top_truncation(default_top_truncation_),
-      value_angle(default_angle_),
-      temperature_systems(std::move(temperature_systems_)),
-      composition_systems(std::move(composition_systems_)),
-      grains_systems(std::move(grains_systems_))
-    {
-      this->type_name = Types::type::Segment;
-
-    }
-
-    template<class A, class B, class C>
-    Segment<A,B,C>::Segment(Segment const &other)
-      :
-      value_length(other.value_length),
-      default_length(other.default_length),
-      value_thickness(other.value_thickness),
-      value_top_truncation(other.value_top_truncation),
-      value_angle(other.value_angle),
-      temperature_systems(other.temperature_systems),
-      composition_systems(other.composition_systems),
-      grains_systems(other.grains_systems)
-    {
-      this->type_name = Types::type::Segment;
-    }
-
-    template<class A, class B, class C>
-    Segment<A,B,C>::~Segment ()
-      = default;
-
-    template<class A, class B, class C>
-    void
-    Segment<A,B,C>::write_schema(Parameters & /*prm*/,
-                                 const std::string & /*name*/,
-                                 const std::string & /*documentation*/) const
-    {
-      WBAssertThrow(false, "not implemented.");
-    }
-
-
-    /**
-    * Todo: Returns a vector of pointers to the Point<3> Type based on the provided name.
-    * Note that the variable with this name has to be loaded before this function is called.
-    */
-    template class
-    Segment<Features::SubductingPlateModels::Temperature::Interface,Features::SubductingPlateModels::Composition::Interface,Features::SubductingPlateModels::Grains::Interface>;
-
-    /**
-    * Todo: Returns a vector of pointers to the Point<3> Type based on the provided name.
-    * Note that the variable with this name has to be loaded before this function is called.
-    */
-    template class
-    Segment<Features::FaultModels::Temperature::Interface,Features::FaultModels::Composition::Interface,Features::FaultModels::Grains::Interface>;
-
-    /**
-    * Todo: Returns a vector of pointers to the Point<3> Type based on the provided name.
-    * Note that the variable with this name has to be loaded before this function is called.
-    */
-    //template class
-    //Segment<char,char>;
-
-  } // namespace Objects
 } // namespace WorldBuilder
